move PDF generator to utils
